Add files via upload
We live in an era of rapid change whereby the dynamic, highly competitive business environment, along with everchanging customer preferences, and the constant emergence of new technologies force organizations to continuously reorganize and reinvent themselves. Research suggests that service organizations are increasingly turning to technological innovations. One such innovation is the Automated Restaurant Table Management (ARTM).

The traditional restaurant model is time consuming and tedious process where a person from the restaurant takes customer name and waits the customer till the tables are unoccupied. The problem arises when the customer queue is large and customer take to occupy seats without the notice of representative. Hence ARTM is used to fully automate the process.

The proposed model aims to improve the ease of use for customers in a restaurant setting. By providing both hardware and software techniques the process makes it streamlined and eliminates need for physical interaction between the customer and representative and traditional way of seating. The use of ARTM makes convenience and ease use of customers. Overall, the proposed model works by prioritize of the customers present in database queue. It works by FIFO (First In First Out) process.

Hardware Setup: The ARTM uses RaspberryPi Cam module to detect table availability using QR (Quick Response) code. There are certain measures to implement Hardware Setup. Proper placement and position of the Camera module is required to detect QR code and a reliable communication is made between the database and the RaspberryPi . 

Software Integration: The ARTM uses software technologies such as HTML (HyperText Markup Language), CSS (Cascading Style Sheets) and JavaScript to create form to store customer data entry and accessed through database. Google Spreadsheet and Cloud Platform is used to store and accessing form data. Python script is used to manage the customer queue and update the database based on output of Hardware Setup. Firebase is used for displaying the final output.